PostToYoutube/requirements.txt
requests==2.28.2
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
